Move the CLI types to internal/
